SEQNG-233 First attempt to implement GMOS N and S.
