+++ modules/edu.gemini.seqexec.server/src/main/scala/edu/gemini/seqexec/server/GmosSouth.scala
@@ -1,131 +1,40 @@
 package edu.gemini.seqexec.server
 
-import java.util.logging.Logger
-
-import edu.gemini.seqexec.model.dhs.ImageFileId
-import edu.gemini.spModel.config2.Config
 import edu.gemini.seqexec.server.ConfigUtilOps._
-import edu.gemini.spModel.gemini.gmos.InstGmosSouth.INSTRUMENT_NAME_PROP
-import edu.gemini.spModel.seqcomp.SeqConfigNames.INSTRUMENT_KEY
-import edu.gemini.spModel.seqcomp.SeqConfigNames.OBSERVE_KEY
+import edu.gemini.seqexec.server.GmosController.{CustomMaskFPU, GmosFPU, UnknownFPU}
+import edu.gemini.spModel.config2.Config
 import edu.gemini.spModel.gemini.gmos.GmosCommonType._
+import edu.gemini.spModel.gemini.gmos.GmosSouthType
 import edu.gemini.spModel.gemini.gmos.GmosSouthType.FPUnitSouth._
-import edu.gemini.spModel.gemini.gmos.InstGmosCommon._
 import edu.gemini.spModel.gemini.gmos.InstGmosSouth._
-import edu.gemini.spModel.obscomp.InstConstants.{INSTRUMENT_NAME_PROP => _, EXPOSURE_TIME_PROP => _, _}
+import edu.gemini.spModel.seqcomp.SeqConfigNames.INSTRUMENT_KEY
+import edu.gemini.spModel.gemini.gmos.InstGmosCommon.{FPU_PROP_NAME, STAGE_MODE_PROP}
 
-import squants.space.LengthConversions._
-import scalaz.{EitherT, Reader, \/}
-import scalaz.syntax.std.string._
-import scalaz.concurrent.Task
+import scalaz.\/
 
-import scala.concurrent.duration._
+final class GmosSouth(val c: GmosSouthController) extends Gmos[GmosSouthController](c) {
+  override val name: String = "GMOS-S"
+  override val dhsInstrumentName: String = "GMOS-S"
 
-final case class GmosSouth(controller: GmosSouthController) extends Instrument {
-  import GmosSouth._
-
-  override val name: String = INSTRUMENT_NAME_PROP
-
-  override val sfName: String = "gmos"
-
-  override val contributorName = "gmosdc"
-  val dhsInstrumentName = "GMOS-S"
-
-  val Log = Logger.getLogger(getClass.getName)
-
-  override def observe(config: Config): SeqObserve[ImageFileId, ObserveResult] = Reader {
-    fileId => controller.observe(fileId).map(_ => ObserveResult(fileId))
-  }
-
-  override def configure(config: Config): SeqAction[ConfigResult] =
-    fromSequenceConfig(config).flatMap(controller.applyConfig).map(_ => ConfigResult(this))
-}
-
-object GmosSouth {
-  val name: String = INSTRUMENT_NAME_PROP
-
-  import GmosSouthController._
-
-  def fpuFromFPUnit(n: Option[FPU], m: Option[String])(fpu: FPUnitMode): GmosFPU = fpu match {
-    case FPUnitMode.BUILTIN     => BuiltInFPU(n.getOrElse(FPU_NONE))
+  override protected def fpuFromFPUnit(n: Option[controller.FPU], m: Option[String])(fpu: FPUnitMode): GmosFPU = fpu match {
+    case FPUnitMode.BUILTIN     => c.BuiltInFPU(n.getOrElse(FPU_NONE))
     case FPUnitMode.CUSTOM_MASK => m match {
       case Some(u) => CustomMaskFPU(u)
       case _       => UnknownFPU
     }
   }
 
-  // It seems this is unused but it shows up on the DC apply config
-  private def biasTimeObserveType(observeType: String): BiasTime = observeType match {
-    case SCIENCE_OBSERVE_TYPE => BiasTimeUnset
-    case FLAT_OBSERVE_TYPE    => BiasTimeUnset
-    case ARC_OBSERVE_TYPE     => BiasTimeEmpty
-    case DARK_OBSERVE_TYPE    => BiasTimeEmpty
-    case BIAS_OBSERVE_TYPE    => BiasTimeUnset
-    case _                    => BiasTimeUnset
-  }
+  override protected def extractFilter(config: Config): \/[ConfigUtilOps.ExtractFailure, GmosSouthType.FilterSouth] = config.extract(INSTRUMENT_KEY / FILTER_PROP).as[GmosSouthType.FilterSouth]
 
-  private def shutterStateObserveType(observeType: String): ShutterState = observeType match {
-    case SCIENCE_OBSERVE_TYPE => UnsetShutter
-    case FLAT_OBSERVE_TYPE    => UnsetShutter
-    case ARC_OBSERVE_TYPE     => OpenShutter
-    case DARK_OBSERVE_TYPE    => CloseShutter
-    case BIAS_OBSERVE_TYPE    => CloseShutter
-    case _                    => UnsetShutter
-  }
+  override protected def extractDisperser(config: Config): \/[ExtractFailure, GmosSouthType.DisperserSouth] = config.extract(INSTRUMENT_KEY / DISPERSER_PROP).as[GmosSouthType.DisperserSouth]
 
-  private def customROIs(config: Config): List[ROI] = {
-    def attemptROI(i: Int): Option[ROI] =
-      (for {
-        xStart <- config.extract(INSTRUMENT_KEY / s"customROI${i}Xmin").as[java.lang.Integer].map(_.toInt)
-        xRange <- config.extract(INSTRUMENT_KEY / s"customROI${i}Xrange").as[java.lang.Integer].map(_.toInt)
-        yStart <- config.extract(INSTRUMENT_KEY / s"customROI${i}Ymin").as[java.lang.Integer].map(_.toInt)
-        yRange <- config.extract(INSTRUMENT_KEY / s"customROI${i}Yrange").as[java.lang.Integer].map(_.toInt)
-      } yield new ROI(xStart, yStart, xRange, yRange)).toOption
+  override protected def extractFPU(config: Config): \/[ExtractFailure, GmosSouthType.FPUnitSouth] = config.extract(INSTRUMENT_KEY / FPU_PROP_NAME).as[controller.FPU]
 
-    val rois = for {
-      i <- 1 to 5
-    } yield attemptROI(i)
-    rois.toList.flatten
-  }
+  override protected def extractStageMode(config: Config): \/[ExtractFailure, GmosSouthType.StageModeSouth] = config.extract(INSTRUMENT_KEY / STAGE_MODE_PROP).as[controller.GmosStageMode]
+}
 
-  def ccConfigFromSequenceConfig(config: Config): TrySeq[CCConfig] =
-    (for {
-      filter           <- config.extract(INSTRUMENT_KEY / FILTER_PROP).as[Filter]
-      disp             <- config.extract(INSTRUMENT_KEY / DISPERSER_PROP).as[Disperser]
-      disperserOrder   =  config.extract(INSTRUMENT_KEY / DISPERSER_ORDER_PROP).as[DisperserOrder]
-      disperserLambda  =  config.extract(INSTRUMENT_KEY / DISPERSER_LAMBDA_PROP).as[java.lang.Double].map(_.toDouble.nanometers)
-      fpuName          =  config.extract(INSTRUMENT_KEY / FPU_PROP_NAME).as[FPU]
-      fpuMask          =  config.extract(INSTRUMENT_KEY / FPU_MASK_PROP).as[String]
-      fpu              <- config.extract(INSTRUMENT_KEY / FPU_MODE_PROP).as[FPUnitMode].map(fpuFromFPUnit(fpuName.toOption, fpuMask.toOption))
-      stageMode        <- config.extract(INSTRUMENT_KEY / STAGE_MODE_PROP).as[GmosStageMode]
-      dtax             <- config.extract(INSTRUMENT_KEY / DTAX_OFFSET_PROP).as[DTAX]
-      adc              <- config.extract(INSTRUMENT_KEY / ADC_PROP).as[ADC]
-      electronicOffset =  config.extract(INSTRUMENT_KEY / USE_ELECTRONIC_OFFSETTING_PROP).as[UseElectronicOffset]
-      disperser = GmosDisperser(disp, disperserOrder.toOption, disperserLambda.toOption)
-    } yield CCConfig(filter, disperser, fpu, stageMode, dtax, adc, electronicOffset.toOption)).leftMap(e => SeqexecFailure.Unexpected(ConfigUtilOps.explain(e)))
+object GmosSouth {
+  val name: String = INSTRUMENT_NAME_PROP
 
-  def dcConfigFromSequenceConfig(config: Config): TrySeq[DCConfig] =
-    (for {
-      obsType      <- config.extract(OBSERVE_KEY / OBSERVE_TYPE_PROP).as[String]
-      biasTime     <- \/.right(biasTimeObserveType(obsType))
-      shutterState <- \/.right(shutterStateObserveType(obsType))
-      exposureTime <- config.extract(OBSERVE_KEY / EXPOSURE_TIME_PROP).as[java.lang.Double].map(_.toDouble.seconds)
-      ampReadMode  <- config.extract(AmpReadMode.KEY).as[AmpReadMode]
-      gainChoice   <- config.extract(INSTRUMENT_KEY / AMP_GAIN_CHOICE_PROP).as[AmpGain]
-      ampCount     <- config.extract(INSTRUMENT_KEY / AMP_COUNT_PROP).as[AmpCount]
-      gainSetting  <- config.extract(INSTRUMENT_KEY / AMP_GAIN_SETTING_PROP).as[String].flatMap(_.parseDouble.disjunction.leftMap(_ => ConversionError(INSTRUMENT_KEY / AMP_GAIN_SETTING_PROP, "Bad Amp gain setting")))
-      xBinning     <- config.extract(INSTRUMENT_KEY / CCD_X_BIN_PROP).as[Binning]
-      yBinning     <- config.extract(INSTRUMENT_KEY / CCD_Y_BIN_PROP).as[Binning]
-      builtInROI   <- config.extract(INSTRUMENT_KEY / BUILTIN_ROI_PROP).as[BuiltinROI]
-      customROI = if (builtInROI == BuiltinROI.CUSTOM) customROIs(config) else Nil
-      roi          <- RegionsOfInterest.fromOCS(builtInROI, customROI).leftMap(e => ContentError(SeqexecFailure.explain(e)))
-    } yield
-      DCConfig(exposureTime, biasTime, shutterState, CCDReadout(ampReadMode, gainChoice, ampCount, gainSetting), CCDBinning(xBinning, yBinning), roi))
-        .leftMap(e => SeqexecFailure.Unexpected(ConfigUtilOps.explain(e)))
-
-  def fromSequenceConfig(config: Config): SeqAction[GmosSouthConfig] = EitherT( Task ( for {
-      cc <- ccConfigFromSequenceConfig(config)
-      dc <- dcConfigFromSequenceConfig(config)
-    } yield GmosSouthConfig(cc, dc)
-  ) )
- }
+  def apply(c: GmosSouthController): GmosSouth = new GmosSouth(c)
+}
