+++ modules/edu.gemini.seqexec.server/src/main/scala/edu/gemini/seqexec/server/SeqexecEngine.scala
@@ -47,7 +47,7 @@ class SeqexecEngine(settings: SeqexecEngine.Settings) {
       if (settings.instForceError) Flamingos2ControllerSimBad
       else Flamingos2ControllerSim
     } else Flamingos2ControllerEpics,
-    if (settings.instSim) GmosSouthControllerSim else GmosControllerEpics
+    if (settings.instSim) GmosSouthControllerSim else GmosSouthControllerEpics
   )
 
   val translatorSettings = SeqTranslate.Settings(
